backend/requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
fastapi[all] >=0.110.0, <0.111.0
honcho >=1.1.0, <1.2.0
psycopg2 >=2.9.5, <2.10.0
pytest >=7.2.1, <7.3.0
pytest-cov >=4.1.0, <4.2.0
python-dotenv >=1.0.0, <1.1.0
alembic >=1.10.2, <1.11.0
pygithub >=1.58.0, <1.59.0
black >=23.10.1, <23.11.0
